+++ src/model/new_line_item.rs
@@ -3,13 +3,13 @@ use clap::Args;
 #[derive(Args)]
 pub struct NewLineItem {
     description: String,
-    quantity: f32,
-    unit_price: f32,
+    quantity: f64,
+    unit_price: f64,
     invoice_id: String,
 }
 
 impl NewLineItem {
-    pub(crate) fn new(description: String, quantity: f32, price: f32, invoice_id: String) -> Self {
+    pub(crate) fn new(description: String, quantity: f64, price: f64, invoice_id: String) -> Self {
         Self {
             description,
             quantity,
@@ -22,11 +22,11 @@ impl NewLineItem {
         &self.description
     }
 
-    pub fn get_quantity(&self) -> f32 {
+    pub fn get_quantity(&self) -> f64 {
         self.quantity
     }
 
-    pub fn get_unit_price(&self) -> f32 {
+    pub fn get_unit_price(&self) -> f64 {
         self.unit_price
     }
